refactor(core): reuse element first create pass in ComponentRef
Reuse element first create pass in ComponentRef.
