+++ packages/core/test/bundling/animations-standalone/bundle.golden_symbols.json
@@ -275,6 +275,7 @@
   "detectChangesInternal",
   "diPublicInInjector",
   "documentElement",
+  "elementStartFirstCreatePass",
   "enterDI",
   "enterView",
   "eraseStyles",
@@ -321,7 +322,6 @@
   "getNullInjector",
   "getOrCreateInjectable",
   "getOrCreateNodeInjectorForNode",
-  "getOrCreateTNode",
   "getOrSetDefaultValue",
   "getOwnDefinition",
   "getParentElement",
@@ -344,7 +344,6 @@
   "importProvidersFrom",
   "includeViewProviders",
   "incrementInitPhaseFlags",
-  "initializeDirectives",
   "inject",
   "injectArgs",
   "injectDestroyRef",
@@ -444,7 +443,6 @@
   "requiresRefreshOrTraversal",
   "resetPreOrderHookFlags",
   "resolveForwardRef",
-  "resolveHostDirectives",
   "resolveTiming",
   "resolveTimingValue",
   "roundOffset",
