+++ packages/core/test/bundling/animations/bundle.golden_symbols.json
@@ -296,6 +296,7 @@
   "detectChangesInternal",
   "diPublicInInjector",
   "documentElement",
+  "elementStartFirstCreatePass",
   "enterDI",
   "enterView",
   "eraseStyles",
@@ -343,7 +344,6 @@
   "getNullInjector",
   "getOrCreateInjectable",
   "getOrCreateNodeInjectorForNode",
-  "getOrCreateTNode",
   "getOrSetDefaultValue",
   "getOwnDefinition",
   "getParentElement",
@@ -367,7 +367,6 @@
   "importProvidersFrom",
   "includeViewProviders",
   "incrementInitPhaseFlags",
-  "initializeDirectives",
   "inject",
   "injectArgs",
   "injectDestroyRef",
@@ -470,7 +469,6 @@
   "requiresRefreshOrTraversal",
   "resetPreOrderHookFlags",
   "resolveForwardRef",
-  "resolveHostDirectives",
   "resolveTiming",
   "resolveTimingValue",
   "roundOffset",
