+++ packages/core/test/bundling/cyclic_import/bundle.golden_symbols.json
@@ -227,6 +227,7 @@
   "detectChangesInViewIfRequired",
   "detectChangesInternal",
   "diPublicInInjector",
+  "elementStartFirstCreatePass",
   "enterDI",
   "enterView",
   "errorContext",
@@ -292,7 +293,6 @@
   "importProvidersFrom",
   "includeViewProviders",
   "incrementInitPhaseFlags",
-  "initializeDirectives",
   "inject",
   "injectArgs",
   "injectDestroyRef",
@@ -380,7 +380,6 @@
   "requiresRefreshOrTraversal",
   "resetPreOrderHookFlags",
   "resolveForwardRef",
-  "resolveHostDirectives",
   "runEffectsInView",
   "runInInjectionContext",
   "saveNameToExportMap",
