+++ packages/core/test/bundling/defer/bundle.golden_symbols.json
@@ -276,6 +276,7 @@
   "detectChangesInViewIfRequired",
   "detectChangesInternal",
   "diPublicInInjector",
+  "elementStartFirstCreatePass",
   "enterDI",
   "enterView",
   "epoch",
@@ -479,6 +480,7 @@
   "init_element_container",
   "init_element_ref",
   "init_element_validation",
+  "init_elements",
   "init_empty",
   "init_environment",
   "init_environment2",
@@ -745,7 +747,6 @@
   "init_zone",
   "init_zoneless_scheduling",
   "init_zoneless_scheduling_impl",
-  "initializeDirectives",
   "inject",
   "injectArgs",
   "injectDestroyRef",
@@ -840,7 +841,6 @@
   "resetPreOrderHookFlags",
   "resolveDirectives",
   "resolveForwardRef",
-  "resolveHostDirectives",
   "retrieveHydrationInfo",
   "runEffectsInView",
   "runInInjectionContext",
