+++ packages/core/test/bundling/forms_reactive/bundle.golden_symbols.json
@@ -329,6 +329,7 @@
   "detectChangesInViewIfRequired",
   "detectChangesInternal",
   "diPublicInInjector",
+  "elementStartFirstCreatePass",
   "enterDI",
   "enterView",
   "epoch",
@@ -429,7 +430,6 @@
   "inheritHostBindings",
   "inheritViewQuery",
   "initFeatures",
-  "initializeDirectives",
   "inject",
   "injectArgs",
   "injectDestroyRef",
@@ -570,7 +570,6 @@
   "resetPreOrderHookFlags",
   "resolveDirectives",
   "resolveForwardRef",
-  "resolveHostDirectives",
   "resolveProvider",
   "runEffectsInView",
   "runInInjectionContext",
