+++ packages/core/test/bundling/forms_template_driven/bundle.golden_symbols.json
@@ -318,6 +318,7 @@
   "detectChangesInternal",
   "diPublicInInjector",
   "elementPropertyInternal",
+  "elementStartFirstCreatePass",
   "enterDI",
   "enterView",
   "epoch",
@@ -415,7 +416,6 @@
   "inheritHostBindings",
   "inheritViewQuery",
   "initFeatures",
-  "initializeDirectives",
   "inject",
   "injectArgs",
   "injectChangeDetectorRef",
@@ -561,7 +561,6 @@
   "resetPreOrderHookFlags",
   "resolveDirectives",
   "resolveForwardRef",
-  "resolveHostDirectives",
   "resolveProvider",
   "resolvedPromise",
   "resolvedPromise2",
