+++ packages/core/test/bundling/hello_world/bundle.golden_symbols.json
@@ -147,6 +147,7 @@
   "cleanUpView",
   "collectNativeNodes",
   "collectNativeNodesInLContainer",
+  "computeStaticStyling",
   "concatStringsWithSpace",
   "config",
   "configureViewWithDirective",
@@ -193,6 +194,7 @@
   "getClosureSafeProperty",
   "getComponentDef",
   "getComponentLViewByIndex",
+  "getConstant",
   "getCurrentTNode",
   "getCurrentTNodePlaceholderOk",
   "getDeclarationTNode",
