+++ packages/core/test/bundling/hydration/bundle.golden_symbols.json
@@ -200,6 +200,7 @@
   "clearElementContents",
   "collectNativeNodes",
   "collectNativeNodesInLContainer",
+  "computeStaticStyling",
   "concatStringsWithSpace",
   "config",
   "configureViewWithDirective",
@@ -254,6 +255,7 @@
   "getClosureSafeProperty",
   "getComponentDef",
   "getComponentLViewByIndex",
+  "getConstant",
   "getCurrentTNode",
   "getCurrentTNodePlaceholderOk",
   "getDOM",
