+++ packages/core/test/bundling/router/bundle.golden_symbols.json
@@ -384,6 +384,7 @@
   "detectChangesInViewIfRequired",
   "detectChangesInternal",
   "diPublicInInjector",
+  "elementStartFirstCreatePass",
   "emptyPathMatch",
   "encodeUriQuery",
   "encodeUriSegment",
@@ -501,7 +502,6 @@
   "includeViewProviders",
   "incrementInitPhaseFlags",
   "initFeatures",
-  "initializeDirectives",
   "inject",
   "injectArgs",
   "injectChangeDetectorRef",
@@ -650,7 +650,6 @@
   "requiresRefreshOrTraversal",
   "resetPreOrderHookFlags",
   "resolveForwardRef",
-  "resolveHostDirectives",
   "rootRoute",
   "routes",
   "runEffectsInView",
