+++ packages/core/test/bundling/standalone_bootstrap/bundle.golden_symbols.json
@@ -173,6 +173,7 @@
   "cleanUpView",
   "collectNativeNodes",
   "collectNativeNodesInLContainer",
+  "computeStaticStyling",
   "concatStringsWithSpace",
   "config",
   "configureViewWithDirective",
@@ -221,6 +222,7 @@
   "getClosureSafeProperty",
   "getComponentDef",
   "getComponentLViewByIndex",
+  "getConstant",
   "getCurrentTNode",
   "getCurrentTNodePlaceholderOk",
   "getDOM",
