+++ packages/core/test/bundling/todo/bundle.golden_symbols.json
@@ -269,6 +269,7 @@
   "detectChangesInViewIfRequired",
   "detectChangesInternal",
   "diPublicInInjector",
+  "elementStartFirstCreatePass",
   "enterDI",
   "enterView",
   "errorContext",
@@ -351,7 +352,6 @@
   "includeViewProviders",
   "incrementInitPhaseFlags",
   "initFeatures",
-  "initializeDirectives",
   "inject",
   "injectArgs",
   "injectDestroyRef",
@@ -453,7 +453,6 @@
   "resetPreOrderHookFlags",
   "resolveDirectives",
   "resolveForwardRef",
-  "resolveHostDirectives",
   "runEffectsInView",
   "runInInjectionContext",
   "saveNameToExportMap",
